identity: Fix manager refcounts, reduce churn
Use identitymanager.Add instead of identitymanager.RemoveOldAddNew
when initially adding the identity of a new endpoint into the identity
manager. This fixes a refcounting bug, where upon the restoration of
two endpoints with the same labels the first restoration adds the
identity to the manager with refcount 1, but the second restoration
first deletes the lone refcount, triggering removal of the associated
policy from the policy cache, and then adds the same ID back, with the
resulting refcount of 1, even when two endpoints are using the same
ID.

Require the old identity given to RemoveOldAddNew to have been added
to the identity manager previously, and make the case where
RemoveOldAddNew is called with two identities with the same numerical
ID a no-op to reduce churn in the observers.

ParseEndpoint(), which is used to restore endpoints on restart, used
to call SetIdentity() to set a nil security identity to the
init-identity.  This is changed so that the caller is responsible for
calling SetIdentity() in all cases.

Fixes: #8685

+++ pkg/endpoint/policy.go
@@ -579,7 +579,7 @@ func (e *Endpoint) runIPIdentitySync(endpointIP addressing.CiliumIP) {
 // SetIdentity resets endpoint's policy identity to 'id'.
 // Caller triggers policy regeneration if needed.
 // Called with e.Mutex Locked
-func (e *Endpoint) SetIdentity(identity *identityPkg.Identity) {
+func (e *Endpoint) SetIdentity(identity *identityPkg.Identity, newEndpoint bool) {
 
 	// Set a boolean flag to indicate whether the endpoint has been injected by
 	// Istio with a Cilium-compatible sidecar proxy.
@@ -596,7 +596,11 @@ func (e *Endpoint) SetIdentity(identity *identityPkg.Identity) {
 	// Current security identity for endpoint is its old identity - delete its
 	// reference from global identity manager, add add a reference to the new
 	// identity for the endpoint.
-	identitymanager.RemoveOldAddNew(e.SecurityIdentity, identity)
+	if newEndpoint {
+		identitymanager.Add(identity)
+	} else {
+		identitymanager.RemoveOldAddNew(e.SecurityIdentity, identity)
+	}
 	e.SecurityIdentity = identity
 	e.replaceIdentityLabels(identity.Labels)
